pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-ingestion-platform"
version = "0.1.0"
description = "Event Ingestion & Enrichment Platform backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "structlog>=25.1.0,<26.0.0",
  "sqlalchemy>=2.0.38,<3.0.0",
  "psycopg[binary]>=3.2.5,<4.0.0",
  "alembic>=1.14.1,<2.0.0",
  "redis>=5.2.1,<6.0.0",
  "celery>=5.5.0,<6.0.0",
  "ua-parser>=1.0.1,<2.0.0",
  "tldextract>=5.3.0,<6.0.0",
  "maxminddb>=2.6.3,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5,<9.0.0",
  "httpx>=0.28.1,<1.0.0",
  "ruff>=0.11.0,<1.0.0",
  "mypy>=1.15.0,<2.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
pretty = true
